move functions from common to location.py
