+++ flowsa/data_source_scripts/BEA.py
@@ -9,7 +9,7 @@ Generation of BEA Gross Output data as FBA, csv files for BEA data
 originate from USEEIOR and were originally generated on 2020-07-14.
 """
 import pandas as pd
-from flowsa.common import US_FIPS
+from flowsa.location import US_FIPS
 from flowsa.settings import externaldatapath
 from flowsa.flowbyfunctions import assign_fips_location_system
