+++ flowsa/data_source_scripts/Blackhurst_IO.py
@@ -11,7 +11,7 @@ import io
 import tabula
 import pandas as pd
 import numpy as np
-from flowsa.common import US_FIPS
+from flowsa.location import US_FIPS
 from flowsa.flowbyfunctions import assign_fips_location_system, \
     load_fba_w_standardized_units
 from flowsa.allocation import \
