+++ flowsa/data_source_scripts/EIA_MECS.py
@@ -11,7 +11,8 @@ Last updated: 8 Sept. 2020
 import io
 import pandas as pd
 import numpy as np
-from flowsa.common import US_FIPS, WITHDRAWN_KEYWORD
+from flowsa.location import US_FIPS
+from flowsa.common import WITHDRAWN_KEYWORD
 from flowsa.settings import vLogDetailed
 from flowsa.flowbyfunctions import assign_fips_location_system
 from flowsa.dataclean import replace_strings_with_NoneType, \
@@ -341,7 +342,7 @@ def eia_mecs_energy_parse(*, df_list, source, year, **_):
     :return: df, parsed and partially formatted to flowbyactivity
         specifications
     """
-    from flowsa.common import assign_census_regions
+    from flowsa.location import assign_census_regions
 
     # concatenate dataframe list into single dataframe
     df = pd.concat(df_list, sort=True)
