+++ flowsa/data_source_scripts/USDA_ACUP.py
@@ -7,7 +7,8 @@ Functions to import and parse USDA Chemical Use Survey
 """
 import json
 import pandas as pd
-from flowsa.common import US_FIPS, WITHDRAWN_KEYWORD, abbrev_us_state, log
+from flowsa.location import US_FIPS, abbrev_us_state
+from flowsa.common import WITHDRAWN_KEYWORD, log
 from flowsa.flowbyfunctions import assign_fips_location_system
