+++ flowsa/data_source_scripts/USDA_CoA_Cropland_NAICS.py
@@ -10,7 +10,8 @@ in NAICS format
 import json
 import numpy as np
 import pandas as pd
-from flowsa.common import WITHDRAWN_KEYWORD, US_FIPS, abbrev_us_state, \
+from flowsa.location import US_FIPS, abbrev_us_state
+from flowsa.common import WITHDRAWN_KEYWORD, \
     fba_wsec_default_grouping_fields
 from flowsa.flowbyfunctions import assign_fips_location_system, \
     equally_allocate_suppressed_parent_to_child_naics
