+++ flowsa/data_source_scripts/USDA_ERS_MLU.py
@@ -11,7 +11,7 @@ Last updated: Thursday, April 16, 2020
 import io
 import pandas as pd
 import numpy as np
-from flowsa.common import get_all_state_FIPS_2, US_FIPS
+from flowsa.location import get_all_state_FIPS_2, US_FIPS
 from flowsa.settings import vLogDetailed
 from flowsa.flowbyfunctions import assign_fips_location_system
 from flowsa.common import load_crosswalk
