+++ flowsa/data_source_scripts/USDA_IWMS.py
@@ -11,7 +11,8 @@ import json
 import pandas as pd
 import numpy as np
 from esupy.dqi import get_weighted_average
-from flowsa.common import US_FIPS, WITHDRAWN_KEYWORD, fba_wsec_default_grouping_fields
+from flowsa.location import US_FIPS
+from flowsa.common import WITHDRAWN_KEYWORD, fba_wsec_default_grouping_fields
 from flowsa.dataclean import replace_strings_with_NoneType, replace_NoneType_with_empty_cells
 from flowsa.flowbyfunctions import assign_fips_location_system, \
     load_fba_w_standardized_units
