+++ flowsa/data_source_scripts/USGS_NWIS_WU.py
@@ -9,8 +9,8 @@ Helper functions for importing, parsing, formatting USGS Water Use data
 import io
 import pandas as pd
 import numpy as np
-from flowsa.common import abbrev_us_state, fba_activity_fields,\
-    capitalize_first_letter, US_FIPS
+from flowsa.location import abbrev_us_state, US_FIPS
+from flowsa.common import fba_activity_fields, capitalize_first_letter
 from flowsa.settings import vLogDetailed
 from flowsa.flowbyfunctions import assign_fips_location_system
 from flowsa.validation import compare_df_units, \
